requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.8.1
    # via
    #   discord-extension-interaction
    #   discord-py
aiosignal==1.2.0
    # via
    #   aiohttp
    #   discord-extension-interaction
async-timeout==4.0.2
    # via
    #   aiohttp
    #   discord-extension-interaction
asyncio==3.4.3
    # via discord-extension-interaction
attrs==21.2.0
    # via
    #   aiohttp
    #   discord-extension-interaction
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   discord-extension-interaction
discord-extension-interaction @ git+https://github.com/gunyu1019/discord-extension-interaction
    # via -r requirements.in
discord.py @ git+https://github.com/Rapptz/discord.py
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
    #   discord-extension-interaction
idna==3.3
    # via
    #   discord-extension-interaction
    #   yarl
multidict==5.2.0
    # via
    #   aiohttp
    #   discord-extension-interaction
    #   yarl
pytz==2022.1
    # via -r requirements.in
yarl==1.7.0
    # via
    #   aiohttp
    #   discord-extension-interaction
